Removing old code
A long time ago we were talking about giving an astraeus the ability to
transport air units, but to prevent shinenigans like pelicans loading up
Pelicans we gave every air unit size 2 and astraeus the capacity to
carry 2. The problem was that this broke auto loading so the idea was
abandoned. The code however remained and has caused problems with area
load on the pelicans. With the code removed area load works again.